requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
dashscope==1.20.4
decorator==5.1.1
editdistance==0.8.1
exceptiongroup==1.2.2
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
funasr==1.1.6
gradio==4.42.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.7
importlib_resources==6.4.3
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.0
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
modelscope==1.17.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.7
oss2==2.18.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==5.27.3
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pynndescent==0.5.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-wpe==0.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
ruff==0.6.1
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.4.0
sseclient==0.0.27
starlette==0.38.2
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.6
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
zhipuai==2.1.4.20230814
zipp==3.20.0
